대여정보 및 결제내역 get api 수정(reservations->payment)
