+++ app/reservations/serializers.py
@@ -415,65 +415,3 @@ class ReservationTimeExtensionUpdateSerializer(serializers.Serializer):
 
     def to_representation(self, instance):
         return ReservationSerializer(instance).data
-
-
-class RentalCarSerializer(serializers.ModelSerializer):
-    min_price = serializers.IntegerField(source='carprice.min_price_per_km')
-    max_price = serializers.IntegerField(source='carprice.max_price_per_km')
-
-    class Meta:
-        model = Car
-        fields = (
-            'number',
-            'manufacturer',
-            'min_price',
-            'max_price',
-        )
-
-
-class RentalSerializer(serializers.ModelSerializer):
-    from_when = serializers.DateTimeField(format='%Y-%m-%d %H:%M', default_timezone=KST)
-    to_when = serializers.DateTimeField(format='%Y-%m-%d %H:%M', default_timezone=KST)
-    time = serializers.SerializerMethodField('get_time')
-    car = RentalCarSerializer()
-    address = serializers.CharField(source='car.zone.address')
-
-    class Meta:
-        model = Reservation
-        fields = (
-            'from_when',
-            'to_when',
-            'time',
-            'car',
-            'address',
-        )
-
-    def get_time(self, reservation):
-        time = int(reservation.time())
-        return f'{time}분'
-
-
-class PaymentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Reservation
-        fields = (
-            'rental_credit',
-            'insurance_credit',
-            'extended_rental_credit',
-            'extended_insurance_credit',
-            'total_credit',
-        )
-
-
-class ReservationAlarmSerializer(serializers.Serializer):
-    reservation_id = serializers.IntegerField(source='id')
-    rental = serializers.SerializerMethodField('get_rental')
-    payment = serializers.SerializerMethodField('get_payment')
-
-    def get_rental(self, reservation):
-        serializer = RentalSerializer(instance=reservation)
-        return serializer.data
-
-    def get_payment(self, reservation):
-        serializer = PaymentSerializer(instance=reservation)
-        return serializer.data
